# shared/openapi/shop_openapi.json
openapi: 3.0.3
info:
  title: ShopDjango API
  version: 1.0.0
  description: API for ShopDjango project
paths:
...
  /api/profile/profiles/me/:
    get:
      operationId: profile_profiles_me_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - profile
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ''
    patch:
      operationId: profile_profiles_me_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedProfile'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedProfile'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedProfile'
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
          description: ''
components:
  schemas:
...
    PatchedProfile:
      type: object
      description: Serializer for user profile management.
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          description: User's first name
          maxLength: 150
        last_name:
          type: string
          description: User's last name
          maxLength: 150
        date_of_birth:
          type: string
          format: date
          nullable: true
          description: User's date of birth (optional)
        phone_number:
          type: string
          description: 'Phone number in international format. Supports digits, spaces,
            hyphens, parentheses, and optional ''+'' prefix. Example: +48131012012'
          pattern: ^\+?[\d\s\-\(\)]+$
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          minimum: -2147483648
          maximum: 2147483647
        profile_completed:
          type: boolean
          readOnly: true
          description: Whether profile has all required information for checkout
        display_name:
          type: string
          readOnly: true
        full_name:
          type: string
          readOnly: true
        age:
          type: integer
          readOnly: true
        is_checkout_ready:
          type: boolean
          readOnly: true
        missing_checkout_fields:
          type: array
          items: {}
          readOnly: true
        shipping_address:
          type: object
          additionalProperties: {}
          readOnly: true
        billing_address:
          type: object
          additionalProperties: {}
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was created
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was last updated
...
    Profile:
      type: object
      description: Serializer for user profile management.
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          description: User's first name
          maxLength: 150
        last_name:
          type: string
          description: User's last name
          maxLength: 150
        date_of_birth:
          type: string
          format: date
          nullable: true
          description: User's date of birth (optional)
        phone_number:
          type: string
          description: 'Phone number in international format. Supports digits, spaces,
            hyphens, parentheses, and optional ''+'' prefix. Example: +48131012012'
          pattern: ^\+?[\d\s\-\(\)]+$
        role:
          allOf:
          - $ref: '#/components/schemas/RoleEnum'
          minimum: -2147483648
          maximum: 2147483647
        profile_completed:
          type: boolean
          readOnly: true
          description: Whether profile has all required information for checkout
        display_name:
          type: string
          readOnly: true
        full_name:
          type: string
          readOnly: true
        age:
          type: integer
          readOnly: true
        is_checkout_ready:
          type: boolean
          readOnly: true
        missing_checkout_fields:
          type: array
          items: {}
          readOnly: true
        shipping_address:
          type: object
          additionalProperties: {}
          readOnly: true
        billing_address:
          type: object
          additionalProperties: {}
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was created
        updated_at:
          type: string
          format: date-time
          readOnly: true
          description: Timestamp when the record was last updated
      required:
      - age
      - billing_address
      - created_at
      - display_name
      - full_name
      - id
      - is_checkout_ready
      - missing_checkout_fields
      - profile_completed
      - shipping_address
      - updated_at
...
    RoleEnum:
      enum:
      - 1
      - 2
      - 3
      type: integer
      description: |-
        * `1` - Admin
        * `2` - Employee
        * `3` - User
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"
...